Update contract on image-snip% other-equal-to? method (racket/gui#119)
The contract on the other-equal-to? method accepted only `image-snip%`
instances, which caused a contract violation when comparing an `image-snip%`
with another non-`image-snip%`, instead of returning #f.

I updated the contract to accept a `snip%` object instead (same contract as
the parent class) and check inside the method if the object is an
`image-snip%`

Here is a sample program to test the fix:

    #lang racket
    (require racket/gui pict pict/snip rackunit)
    (define s1 (make-object image-snip% (pict->bitmap (text "hello"))))
    (define s2 (make-object pict-snip% (text "hello")))
    (define s3 (make-object image-snip% (pict->bitmap (text "hello"))))

    (check-true (equal? s1 s1))
    (check-false (equal? s1 s2))
    (check-false (equal? s2 s1)) ; this case used to report a contract violation
    (check-true (equal? s1 s3))
    (check-true (equal? s3 s1))