make 2htdp/image usable within sandbox
Calling `get-preference*` will attempt to read `org.racket-lang.prefs.rktd`
or `racket-prefs.rktd` which might not be granted the read access within
a sandbox, resulting in the error:

  file-or-directory-modify-seconds: `read' access denied for ....

This commit catches the error if this situation happens and default back
to the default value. Also delay the evaluation so that if we only
require the file, the preference doesn't get loaded.